Properly escape libexecdir path configure argument
This puts iwd binary in /usr/lib/<arch-triplet>/iwd/iwd.
Before it ended up in /lib/<arch-triplet>/iwd/iwd because
of the missing escape of the $prefix variable in the
configure argument.

Spotted (and fix suggested) by Michael Biebl,
who also suggested maybe it's better to just use /usr/libexec
(which is the default).