Build against external (lib)ell
The ell library is still not at version 1.0 and thus still don't
have any ABI guarantees (upstream). However at this point in the
release cycle we can hope for no surprises. We'll soon be in
hard-freeze.

(For bullseye development cycle we'll go back to using the bundled
version, waiting for ell to release version 1.0.)

Closes: #911057