Do not enable/start/restart service by default
It is possible to use iwd standalone. People interested in that
should enable the service manually. (i.e. systemctl enable --now iwd)
The assumption is at this point that most people however likely will
rely on the dbus activation method of having iwd started when used.
eg. Triggered by NetworkManager (when using the iwd backend), triggered
by iwctl sending a dbus request, or some other manager.

While at it also avoid stopping the service on upgrades, as this
would cause network outtage on upgrades. People upgrading remotely
will be nervous if this happens in the middle of the upgrade
procedure, however not restarting ever is also bad. Hopefully
people are using things like needrestart etc to detect it.